+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 * A file containing the LOTR biome ids now will be generated on startup (if the LOTR mod is present)
 * Added missing comments to some configuration properties
 * Fixed that hired units from the LOTR Mod start attacking players/units with PvP mode OFF (without causing damage)
+* The PvP mode of players involved in a PvP event cannot be changed
+* Players which are in PvP cannot use the LOTR Mod fast travel system
+* Added a command blacklist for players which are in PvP: While in PvP, they cannot use the blacklisted commands
 
 ## 1.0.0-BETA
 * Added combat logging (two handlers: csv (default) and simple)